+++ src/magicLink.ts
@@ -151,8 +151,9 @@ export abstract class MagicLinkToken extends Token {
 			return errors;
 		}
 		// eslint-disable-next-line @typescript-eslint/no-unused-expressions
-		/[，；。：！？（）]+|\|+/gu;
-		const source = `[，；。：！？（）]+${this.type === 'ext-link-url' ? String.raw`|\|+` : ''}`,
+		/[，；。：！？（）]+/gu;
+		const pipe = this.type === 'ext-link-url',
+			source = pipe ? String.raw`\|+` : '[，；。：！？（）]+',
 			regex = new RegExp(source, 'u'),
 			regexGlobal = new RegExp(source, 'gu');
 		for (const child of this.childNodes) {
@@ -169,8 +170,7 @@ export abstract class MagicLinkToken extends Token {
 					left = lines[top - 1]!.length,
 					startIndex = refError.startIndex + index,
 					startLine = refError.startLine + top - 1,
-					startCol = top === 1 ? refError.startCol + left : left,
-					pipe = s.startsWith('|');
+					startCol = top === 1 ? refError.startCol + left : left;
 				const e = {
 					...refError,
 					message: Parser.msg('$1 in URL', pipe ? '"|"' : 'full-width punctuation'),